do some refactor on build process
