+++ Gruntfile.js
@@ -140,7 +140,7 @@ module.exports = function(grunt) {
         exec: {
             'webpack-dev': `${path.resolve('./node_modules/.bin/webpack')} --config-name dev`,
             'webpack-prod': `${path.resolve('./node_modules/.bin/webpack')} --config-name prod`,
-            'webpack-electron': `${path.resolve('./node_modules/.bin/webpack')} --config-name electron`,
+            'webpack-electron': `${path.resolve('./node_modules/.bin/webpack')} --config webpack.electron.config`,
             'webpack-all': `${path.resolve('./node_modules/.bin/webpack')}`,
         },
         sass: {
@@ -338,14 +338,7 @@ module.exports = function(grunt) {
     grunt.registerTask('build-dev', ['clean:intermediates', 'exec:webpack-dev', 'build-assets', 'drop:dev']);
     grunt.registerTask('build-prod', ['clean:intermediates', 'exec:webpack-prod', 'build-assets', 'drop:production']);
     grunt.registerTask('build-electron', ['clean:intermediates', 'exec:webpack-electron', 'build-assets', 'drop:electron']);
-    grunt.registerTask('build-all', [
-        'clean:intermediates',
-        'exec:webpack-all',
-        'build-assets',
-        'drop:dev',
-        'drop:electron',
-        'release-drops',
-    ]);
+    grunt.registerTask('build-all', ['clean:intermediates', 'exec:webpack-all', 'build-assets', 'drop:dev', 'release-drops']);
 
     grunt.registerTask('default', ['build-dev']);
 };
